+++ metka/src/main/java/fi/uta/fsd/metka/data/entity/impl/StudyEntity.java
@@ -15,19 +15,21 @@ import javax.persistence.*;
  * IMPORTANT: studyNumber is not used for references in any way, the only meaning it has is to provide a sequence of numbers
  *            to provide a new study with a study number that's not affected by other revisionable objects.
  *            Since sequences are not supported with annotations and non ID fields manual sequencing is required instead.
- *            This should be implemented using SequenceEntity table and using ConfigurationType value as a key
+ *            This should be implemented using SequenceEntity table and using ConfigurationType value as a key.
+ *
+ * TODO: This study number field should propably be changed to the actual study number string since that's what users know.
  */
 @Entity
 @DiscriminatorValue(ConfigurationType.Values.STUDY)
 public class StudyEntity extends RevisionableEntity {
     @Column(name = "STUDY_NUMBER", updatable = false, unique = true)
-    private Integer studyNumber;
+    private Long studyNumber;
 
-    public Integer getStudyNumber() {
+    public Long getStudyNumber() {
         return studyNumber;
     }
 
-    public void setStudyNumber(Integer studyNumber) {
+    public void setStudyNumber(Long studyNumber) {
         this.studyNumber = studyNumber;
     }
 }
